requirements.txt
pymongo==4.6.1
pymongo[srv]
certifi
scikit-learn==1.4.0
xgboost==2.0.3
catboost==1.2.3
pandas==2.2.0
numpy==1.26.3
Flask==3.0.1
evidently==0.4.13
pyyaml==6.0.1
dill==0.3.8
shap==0.44.1
imbalanced-learn==0.12.0
boto3==1.34.34
mypy-boto3-s3==1.34.14
python-dotenv==1.0.1
-e .
